feat: write client secrets to local storage
This makes sure that client secrets aren't cleared once the page
session ends, which can be detrimental to UX if the user has to
enter their password every time the page is reloaded.

In the future, implementing a Content Security Policy is absolutely
critical to enhance user security and to prevent secrets from being
tampered with.