fix: decode client_random in server
Since we're using a custom codec for the client_random field, we
need to manually decode the Base64 in the handler before it can
be used to compute the salt. Ideally this shouldn't happen if
a specialized type were returned, which should be solvable once
we figure out how to handle Base64 as a universal type.